requirements.txt
torch
matplotlib
opencv-contrib-python
imutils
scikit-learn
tqdm
pandas
pyarrow
pyyaml
tensorboard
torchvision
pillow
numpy<2.0
focal-frequency-loss
tensorboard
prettytable
pytorch_ssim
torcheval
pytorch_msssim
scikit-image